Initial commit:  Select libs from my research code.
This is a subset of my research code, which I've been porting from a
combination of C and FORTRAN to a combination of C++ and FORTRAN.  Uses
C++03 features & some of the Boost libraries.  The "subset" contains
only a few libraries, ones that have been completely ported over to C++
and are [mostly] standalone.

This code *does* compile ... when built from inside of the workspace
containing my full research-code.  The subset in this commit, however,
may have dependencies that I did not notice.  I'll be checking that
next.